refactor(macos): split AIHandler.swift along inference and Ollama seams
Pulled the inference flow (infer + native-model resolution + page-context
preload + screenshot capture) into AIHandler+Inference.swift and the
Ollama HTTP transport (chat/generate clients + reachability probe + URL
helpers + capability inference) into AIHandler+Ollama.swift. The main
file now owns the type, request dispatch, status/load/unload/record
handlers, and the small error-mapping utilities shared by every backend.

The `AIBackendStore` actor was widened from `private` to module-internal
so the extensions can read its snapshot; it is still scoped to the AI
handler files and not used elsewhere.

before: AIHandler.swift 701 lines
after:
  AIHandler.swift            273
  AIHandler+Inference.swift  282
  AIHandler+Ollama.swift     177
